Add fresh.py as used in training with HUBERT_BASE
Vanilla Prak tries to be easy to understand and therefore uses just a plain ReLU stack.
In "Fresh Prak", we give up these limitations and use whatever the modern times bring.

Lot of related things is in not-yet-committed notebooks so fresh.py by itself
is not of much use yet. We will continue refactorization, first of all to
make the inference working with the 'fresh' (and more precise) model.